+++ CHANGES.md
@@ -1,5 +1,9 @@
 # aws-api 0.8
 
+## DEV
+
+* handle doubles when parsing timestamps [#36](https://github.com/cognitect-labs/aws-api/issues/36)
+
 ## 0.8.198 / 2019-01-11
 
 * support HAL responses [#30](https://github.com/cognitect-labs/aws-api/issues/30)